# lib/ignore: Replace lib/fnmatch with github.com/gobwas/glob