### Purpose

This globbing library is literally *ten times* faster than my fnmatch implementation:

```
benchmark                  old ns/op     new ns/op     delta
BenchmarkMatch-8           13842         1200          -91.33%
BenchmarkMatchCached-8     139           147           +5.76%
    
benchmark                  old allocs     new allocs     delta
BenchmarkMatch-8           0              0              +0.00%
BenchmarkMatchCached-8     0              0              +0.00%
    
benchmark                  old bytes     new bytes     delta
BenchmarkMatch-8           12            0             -100.00%
BenchmarkMatchCached-8     0             0             +0.00%
```

### Testing

Our ignore implementation is actually fairly well covered by unit tests so this doesn't worry me so much.

### Documentation

This is API breakage as we no longer return the compiled patterns as regexps (there are no regexps). Instead we return the original patterns, but expanded as we've done while parsing them and explicitly indicating case insensitiveness (which in my case comes by default as this is on Mac). Example new format:

```
{
  "ignore": [
    "/Backups"
  ],
  "patterns": [
    "(?i)/Backups",
    "(?i)/Backups/**"
  ]
}
```