@calmh 2016-04-02
Hm, there's a little bit more work to do here around Windows file path separators.