@calmh 2016-04-02
So frankly I get a little confused when looking at our current implementation. We look for things like strings starting with "/**" - don't we support using "\" in patterns on Windows? Shouldn't we?