@AudriusButkevicius 2016-04-02
I think the ignores are always supposed to be linux separators