@AudriusButkevicius 2016-04-02
The only problem I see with this is that inotify uses this, and whoever else wants to use this, has to implement the grammar themselves (unless they use go and can use the same library)