@calmh 2016-04-02
I think that at the very least we should support both, on Windows. `//` will need to be a comment no matter what, but both `foo\bar\**` and `foo/bar/**` *should* be valid patterns.