@calmh 2016-04-02
I think the grammar has always been "glob patterns" which doesn't change. That we exposed the regexps were an unfortunate implementation detail that I added for debugging purposes, and that then got used... :/