@AudriusButkevicius 2016-04-02
So the definition of glob patterns is not very defined. I think this library supports c{ow,at} type of globs which others languages might not have.