@canton7 2016-04-02
If you change the format of "patterns" away from regexes, I'll probably crash as .NET tries to compile invalid regexes. If you change the key to something like "glob_patterns", I'll be a lot happier :smile: 

Not sure how much you care about inotify implementations not telling Syncthing about changes to ignored files: if you're planning on integrating that into Syncthing for 0.13 then of course it doesn't matter...