@calmh 2016-04-02
It also exposes a bug in our previous implementation, IMHO. We expect `**/efile` to match `efile` at the top level, which this library does not...