@calmh 2016-04-02
> Not sure how much you care about inotify implementations not telling Syncthing about changes to ignored files

It's pretty much an optimization. We should ignore them anyway and do nothing. We might have a bug though where we do a full scan when asked to scan just an ignored file; need to double check that.