@calmh 2016-04-02
I added a test to verify that we understand both `a/b` and `a\b` on Windows (we're supposed to not handle escapes on Windows, which should mean we understand both kinds of slashes as path separators).