@calmh 2016-04-02
Can someone with a Windows box run 

```
cd lib/ignore
go test -run None -bench . -benchmem
```

on both `master` and this branch? I wonder how painful the slashification is on Windows.