@canton7 2016-04-02
Master:

```
$ go test -run None -bench . -benchmem
PASS
BenchmarkMatch     30000             50419 ns/op               1 B/op          0 allocs/op
BenchmarkMatchCached     5000000               244 ns/op               0 B/op          0 allocs/op
ok      github.com/syncthing/syncthing/lib/ignore       3.585s
```

calmh/glob:
```
$ go test -run None -bench . -benchmem
PASS
BenchmarkMatch    500000              3057 ns/op               0 B/op          0 allocs/op
BenchmarkMatchCached     5000000               242 ns/op               0 B/op          0 allocs/op
ok      github.com/syncthing/syncthing/lib/ignore       3.071s
```