@calmh 2016-04-02
Ok so enormously faster on Windows too. To me that nails it - it's worth the pain of no longer exposing regexps.