@AudriusButkevicius 2016-04-02
So my browser refuses to switch to the code tab for 60 files. Did you rename the key as canton7 asked?