@calmh 2016-04-02
No, I haven't. I will!

As for the diff, all the interesting things are on ignore.go, the rest is just a dependency shuffle.