@calmh 2016-04-02
```diff
--- a/cmd/syncthing/gui.go
+++ b/cmd/syncthing/gui.go
@@ -933,8 +933,8 @@ func (s *apiService) getDBIgnores(w http.ResponseWriter, r *http.Request) {
        }
 
        sendJSON(w, map[string][]string{
                "ignore":   ignores,
-               "patterns": patterns,
+               "parsed": patterns,
        })
 }
 ```