@calmh 2016-04-02
Does github... duplicate every line break in the commit message? Wtf?